Added error checking
Script file now checks for auto-named Macros and Macro Groups and handles them gracefully. It'll try and name a Macro based on the name of the first action, and if there's no name for the first action it'll name it by the action type instead. All "Smart Macro Groups" are skipped (instead of just the two default ones), and "Untitled Macro Groups" (auto-named and missing the 'name' field in the KM plist) are named semi-correctly.